+++ testData/psi/resolve/var/Issue764_included.bash
@@ -1,0 +1,7 @@
+function checkPrerequisites() {
+    if 1; then
+        exit
+    fi
+}
+
+VAR=1